Refactor DDB.
