+++ sources/dynamodb/snapshot/snapshot_iterator.go
@@ -1,4 +1,4 @@
-package dynamodb
+package snapshot
 
 import (
 	"fmt"
